+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 1.11.2
 
+Features:
+* Support xonsh detection (TerminalShell)
+
 Bugfixes:
 * Fix flatpak package count (#441)
 * Don't print white color blocks with `--pipe` (#450)